Jump improvement + state handler + object picking up and throwing
Added a state handler which will handle player sprite state. Also added more gravity on the downward arc of the jump. Also added the abilty to pick objects on contact and throw them with q. Still need to decouple the objects and players to allow other objects to parent from them.