Room moving + coyote time + state handler updating + organization
Room moving is now possible with exit objects. it has minor side effects but works

coyote time is implemented. right now the time is very generous for testing purposes.

state handler got updated to be in the player_master step event with an animation state. individual player objects will need to implement the animation updater.

I generally organized everything to allow for better parenting and making sense of the code